% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_construction.R
\name{fillingInSpace}
\alias{fillingInSpace}
\title{Spatial reconstruction}
\usage{
fillingInSpace(
  lstInput,
  studyarea = "ATA",
  covariable1,
  covariable2,
  nc = 4,
  bs = "cr",
  k = 20
)
}
\arguments{
\item{lstInput}{MODIS LST scene to be filled in space, using GAM with 3D spatial surface trend.}

\item{studyarea}{Name of study area ( 'GADM','countries'), the default study area is Antarctica "ATA".}

\item{covariable1}{Digital elevation model as additional variable to improve the prediction of LST values. It should be a raster object and match the extent and origin of "lstInput" variable.}

\item{covariable2}{Any additional variable to improve the prediction of LST values. It should be a raster object and match the extent and origin of "lstInput" variable.}

\item{nc}{Cluster size (Default = 4).}

\item{bs}{Represent the smooth function and quadratic penalty to control degree of freedom of smoothing (Default value is cubic spline basis "cr").}

\item{k}{Number of knots used for smoothing the function and decide the degree of freedom (Default = 20).}
}
\value{
Filled MODIS LST scene.
}
\description{
Fill the gaps in the MODIS LST scene using GAM with 3D spatial trend surface, and two covariables used  to improve the results. The covariables have to be raster objects and matching the input MODIS LST scene.
}
\examples{

aqua <- raster(system.file("MODIS_data","aqua.tif", package="modislst"))
dem <- raster(system.file("covariables","dem.tif", package="modislst"))
aspect <- raster(system.file("covariables","aspect.tif", package="modislst"))
fillingInSpace(aqua, covariable1=dem, covariable1=aspect)




}
