% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_construction.R
\name{fillingInTime}
\alias{fillingInTime}
\title{Temporal reconstruction}
\usage{
fillingInTime(inputs, m = 47)
}
\arguments{
\item{inputs}{List of length two with following inputs: the first element is the target scene, the second element is the reference scene.}

\item{m}{The size of moving window. Default value is 47, using bigger size would fill more pixels but increase the computing work and reduce the accuracy of filled pixels.}
}
\value{
MODIS LST scene filling partially.
}
\description{
Fill the gaps in the target MODIS LST scene from the nearest MODIS LST scene (reference scene).
}
\examples{
\dontrun{
fillingLST(inputs=list(aqua,terra), m=47)
}

}
