% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_construction.R
\name{fillingInTime}
\alias{fillingInTime}
\title{Temporal reconstruction}
\usage{
fillingInTime(target, reference, m = 47)
}
\arguments{
\item{target}{Target MODIS LST scene to be filled. Raster object in kelvin unit.}

\item{reference}{Reference MODIS LST scene to fill from it. Raster object in kelvin unit.}

\item{m}{The size of moving window. Default value is 47, using bigger size would fill more pixels but increase the computing work and reduce the accuracy of filled pixels.}
}
\value{
MODIS LST scene filling partially.
}
\description{
Fill the gaps in the target MODIS LST scene from the nearest MODIS LST scene (reference scene).
}
\examples{
\dontrun{
aqua <- raster(system.file("MODIS_data","aqua.tif", package="modislst"))
terra <- raster(system.file("MODIS_data","terra.tif", package="modislst"))
fillingLST(inputs=list(aqua,terra), m=47)
}

}
