% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{lstProcessing}
\alias{lstProcessing}
\title{Pre-processing function}
\usage{
lstProcessing(LSTScene, studyarea = "ATA", crs = "+init=epsg:3031")
}
\arguments{
\item{LSTScene}{MODIS LST scene to be cropped and projected.}

\item{studyarea}{Name of study area ( 'GADM','countries'), the default study area is Antarctica "ATA".}

\item{crs}{Target projected coordinate system, the default is Antarctica polar stereographic projected coordinate system "+init=epsg:3031".}
}
\value{
Projected and Cropped MODIS LST scenes.
}
\description{
Crop MODIS LST scenes based on the extent of the study area,after that reproject cropped MODIS LST scenes to the target projected coordinate system.
}
\examples{

lstProcessing(aqua,terra)



}
